+++ packages/components/src/DataTable/stories/Interaction.tsx
@@ -39,7 +39,12 @@ import { filters as defaultFilters } from '../../__mocks__/filters'
 import { columns as mockColumns } from '../../__mocks__/DataTable/columns'
 import { data } from '../../__mocks__/DataTable/data'
 import { DataTableColumns } from '../Column'
-import { actions, itemBuilder } from './items'
+import {
+  actions,
+  itemBuilder,
+  itemsActionsActionPrimary,
+  itemsActionPrimary,
+} from './items'
 
 interface DemoProps extends Omit<DataTableProps, 'bulk' | 'select'> {
   bulk: boolean
@@ -143,6 +148,10 @@ const Template: Story<DemoProps> = ({
   )
 }
 
+const TemplateAction: Story<DataTableProps> = ({ ...args }) => {
+  return <DataTable {...args} />
+}
+
 export const Basic = Template.bind({})
 Basic.args = {
   columns: mockColumns,
@@ -182,4 +191,16 @@ SelectBulkActiveItems.args = {
   ...Select.args,
   bulk: true,
   selectedItems: ['cheddar', 'gouda'],
+}
+
+export const ActionsAndPrimaryAction = TemplateAction.bind({})
+ActionsAndPrimaryAction.args = {
+  children: itemsActionsActionPrimary,
+  columns: mockColumns,
+}
+
+export const PrimaryAction = TemplateAction.bind({})
+PrimaryAction.args = {
+  children: itemsActionPrimary,
+  columns: mockColumns,
 }